Updated 3dsmax plugins to new workflow.
