+++ hooks/tk-multi-publish2/3dsmax.basic/start_version_control.py
@@ -15,9 +15,10 @@ import sgtk
 HookBaseClass = sgtk.get_hook_baseclass()
 
 
-class MaxVersionUpPlugin(HookBaseClass):
+class MaxStartVersionControlPlugin(HookBaseClass):
     """
-    Plugin for creating the next version of a file.
+    Simple plugin to insert a version number into the max file path if one
+    does not exist.
     """
 
     @property
@@ -39,7 +40,7 @@ class MaxVersionUpPlugin(HookBaseClass):
         """
         One line display name describing the plugin
         """
-        return "Save the file to the next version"
+        return "Add version number to Max file name"
 
     @property
     def description(self):
@@ -48,11 +49,9 @@ class MaxVersionUpPlugin(HookBaseClass):
         contain simple html for formatting.
         """
         return """
-        Detect the version number in the max file path and save it to the next
-        available version number. The plugin will only be available to max
-        files where a version number can be detected in the path. The plugin
-        will only create the next available version if the new path does not
-        already exist.
+        This plugin will acts on Maya session files where a version number
+        can not be detected in the file name. If checked, this plugin will
+        insert a version number into the file name and save the session.
         """
 
     @property
@@ -114,43 +113,37 @@ class MaxVersionUpPlugin(HookBaseClass):
         """
 
         publisher = self.parent
-        path = MaxPlus.FileManager.GetFileNameAndPath()
-        checked = True
+        path = _session_path()
 
-        if MaxPlus.FileManager.IsSaveRequired() or not path:
-            # the session has unsaved changes. provide a save button and uncheck
-            # the item. the session will need to be saved before validation will
-            # succeed.
-            if not path:
-                self.logger.warn(
-                    "Unsaved changes in the session",
-                    extra=_get_save_as_action()
-                )
-                checked = False
-        else:
-            # we have a path and there are no unsaved changes. make sure the
-            # path has a version number. we accept it regardless since it can be
-            # saved with a different file name while the publisher is open. but
-            # we will set it to be unchecked by default if there is no version
-
-            # get the path in a normalized state. no trailing separator,
-            # separators are appropriate for current os, no double separators,
-            # etc.
-            path = sgtk.util.ShotgunPath.normalize(path)
+        if path:
             version_number = publisher.util.get_version_number(path)
-
-            if version_number is None:
-                self.logger.warn(
-                    "No version number detected in the file name",
-                    extra=_get_version_docs_action()
-                )
-                checked = False
+            if version_number is not None:
+                self.logger.info(
+                    "Maya '%s' plugin rejected the current Maya session...")
+                self.logger.info(
+                    "  There is already a version number in the file...")
+                self.logger.info("  Maya file path: %s" % (self.name,))
+                return {"accepted": False}
+        else:
+            # the session has not been saved before (no path determined).
+            # provide a save button. the session will need to be saved before
+            # validation will succeed.
+            self.logger.warn(
+                "The Maya session has not been saved.",
+                extra=_get_save_as_action()
+            )
 
         self.logger.info(
-            "3dsMax version up plugin accepted the current Max session.")
+            "Maya '%s' plugin accepted the current Maya session." %
+            (self.name,),
+            extra=_get_version_docs_action()
+        )
+
+        # accept the plugin, but don't force the user to add a version number
+        # (leave it unchecked)
         return {
             "accepted": True,
-            "checked": checked,
+            "checked": False
         }
 
     def validate(self, settings, item):
@@ -167,58 +160,16 @@ class MaxVersionUpPlugin(HookBaseClass):
         :returns: True if item is valid, False otherwise.
         """
 
-        path = MaxPlus.FileManager.GetFileNameAndPath()
-        publisher = self.parent
+        path = _session_path()
 
-        if MaxPlus.FileManager.IsSaveRequired() or not path:
+        if not path:
             # the session still requires saving. provide a save button.
-            # validation fails since we don't want to save as the next version
-            # until the current changes have been saved.
+            # validation fails
             self.logger.error(
-                "Unsaved changes in the session",
+                "The Max session has not been saved.",
                 extra=_get_save_as_action()
             )
             return False
-        else:
-            # we have a path and there are no unsaved changes. make sure the
-            # path has a version number. if not, validation fails
-
-            # get the path in a normalized state. no trailing separator,
-            # separators are appropriate for current os, no double separators,
-            # etc.
-            path = sgtk.util.ShotgunPath.normalize(path)
-
-            version_number = publisher.util.get_version_number(path)
-
-            if version_number is None:
-                # still no version number in the file name. a warning was
-                # provided in accept(). validation fails
-                self.logger.error(
-                    "No version number detected in the file name",
-                    extra=_get_version_docs_action()
-                )
-                return False
-
-        next_version_path = publisher.util.get_next_version_path(path)
-
-        # nothing to do if the next version path can't be determined or if it
-        # already exists.
-        if not next_version_path:
-            self.logger.error("Could not determine the next version path.")
-            return False
-        elif os.path.exists(next_version_path):
-            self.logger.error(
-                "The next version of the path already exists",
-                extra={
-                    "action_show_folder": {
-                        "path": next_version_path
-                    }
-                }
-            )
-            return False
-
-        # insert the path into the properties for use during the publish phase
-        item.properties["next_version_path"] = next_version_path
 
         return True
 
@@ -232,10 +183,22 @@ class MaxVersionUpPlugin(HookBaseClass):
         :param item: Item to process
         """
 
-        # get the next version path and save the document
-        next_version_path = item.properties["next_version_path"]
-        MaxPlus.FileManager.Save(next_version_path)
-        self.logger.info("The 3dsMax session is now at the next version!")
+        publisher = self.parent
+
+        # get the path in a normalized state. no trailing separator, separators
+        # are appropriate for current os, no double separators, etc.
+        path = sgtk.util.ShotgunPath.normalize(_session_path())
+
+        # ensure the session is saved in its current state
+        _save_session(path)
+
+        # get the path to a versioned copy of the file.
+        version_path = publisher.util.get_version_path(path, "v001")
+
+        # save to the new version path
+        _save_session(version_path)
+        self.logger.info("A version number has been added to the Max file...")
+        self.logger.info("  Max file path: %s" % (version_path,))
 
     def finalize(self, settings, item):
         """
@@ -251,13 +214,29 @@ class MaxVersionUpPlugin(HookBaseClass):
         pass
 
 
+def _session_path():
+    """
+    Return the path to the current session
+    :return:
+    """
+
+    return MaxPlus.FileManager.GetFileNameAndPath()
+
+
+def _save_session(path):
+    """
+    Save the current session to the supplied path.
+    """
+    MaxPlus.FileManager.Save(path)
+
+
 def _get_save_as_action():
     """
     Simple helper for returning a log action dict for saving the session
     """
     return {
         "action_button": {
-            "label": "Save",
+            "label": "Save As...",
             "tooltip": "Save the current session",
             "callback": MaxPlus.FileManager.SaveAs
         }
